Add new .gitignore, .css, and .razor files along with updated app styles
This commit includes addition of new .gitignore, .css, and .razor files. It also includes important updates to app.css to implement new styling features, and ensures the structure of the app conforms with the new routes and focus navigations specified in App.razor. Furthermore, a minified version of Bootstrap v5.1.0 was added for faster load times and greater performance.