Add 'About' page and 'Bibliography' component for the 'Memes2' section
New 'About' page has been added with the primary layout in Layout/Memes2/AboutLayout.razor. Alongside this, a new 'Bibliography' component for the 'Memes2' section was introduced in Components/Bib/BibliographyMemes2.razor. A model for bibliography resources was also created to handle book, paper, and website resources. Finally, the main project file was updated to include the new 'Model' folder.