Add PDF Viewer component and Keen manual page
Added a reusable PDF Viewer component and a new Keen Manual page that uses this component. The PDF Viewer will allow PDF files to be displayed in an iframe. The Keen Manual page will display the manual for the Keen library in a PDF format. Also updated the `NavMenuKeen.razor` to include a link for this new Manual page.