Add CleanCode page and modify LessonList
A new page, CleanCode.razor, was added to the Memes2/_Lessons directory. The page uses the NavigationManager to navigate to a specified SharepointResource. Furthermore, Modifications were made to the LessonList.razor file where a new lesson on "Programación Orientada a Objetos V" was added and "Property-based Testing V" and "Property-based Testing VI" were commented out.